refactor utils and onchainConfigCodec to handle felts vs bigInts
